fix: 중복 코드 정리 - 중복 코드 custom hook으로 분리
