+++ src/hooks/usePostList.ts
@@ -1,20 +1,20 @@
-import { useCallback, useState } from 'react';
-import { useRouter } from 'next/navigation';
+import { useState } from 'react';
 import { useAppSelector } from '@/redux/storeConfig';
 import useSearchTutor from '@/hooks/useSearchTutor';
 import { SearchParamsType } from '@/types/posts.types';
 import useRecentSearchHistory from '@/hooks/useRecentSearchHistory';
 import { useQueryClient } from '@tanstack/react-query';
 import { RECENT_SEARCH_KEY } from '@/api/signup/keys.constant';
+import useQueryParams from '@/hooks/useQueryParams';
 
 export const usePostList = (initialSearchParams: SearchParamsType) => {
-  const router = useRouter();
   const [currentPage, setCurrentPage] = useState<number>(1);
   const [tutor, setTutor] = useState<string>('0');
   const [activeTab, setActiveTab] = useState<string>('all');
   const [keyword, setKeyword] = useState<string>(
     initialSearchParams?.title || '',
   );
+  const updateQueryParams = useQueryParams();
 
   const { recentSearchMutation } = useRecentSearchHistory();
   const queryClient = useQueryClient();
@@ -23,35 +23,16 @@ export const usePostList = (initialSearchParams: SearchParamsType) => {
     (state) => state.authSlice,
   );
 
-  const updateQueryParams = useCallback(
-    (key: string, value: string | number | undefined) => {
-      const query = new URLSearchParams(window.location.search);
-      if (value && !(key === 'tutorId' && value === 'all')) {
-        query.set(key, String(value));
-      } else {
-        query.delete(key);
-      }
-
-      if (key !== 'page') {
-        setCurrentPage(1);
-        query.set('page', '1');
-      }
-
-      router.push(`/?${query.toString()}`);
-    },
-    [router],
-  );
-
   const getFetchTutors = useSearchTutor(trackName, loginPeriod, loginPeriod);
 
   const handleTabChange = (newTab: string) => {
     setActiveTab(newTab);
-    updateQueryParams('postType', newTab);
+    updateQueryParams('postType', newTab, setCurrentPage);
   };
 
   const handlePageChange = (page: number) => {
     setCurrentPage(page);
-    updateQueryParams('page', page);
+    updateQueryParams('page', page, setCurrentPage);
   };
 
   const handleSearchClick = (
@@ -60,7 +41,7 @@ export const usePostList = (initialSearchParams: SearchParamsType) => {
   ) => {
     e.stopPropagation();
     setKeyword(keyword);
-    updateQueryParams('title', keyword);
+    updateQueryParams('title', keyword, setCurrentPage);
   };
 
   const handleKeywordSearch = async () => {
@@ -76,12 +57,12 @@ export const usePostList = (initialSearchParams: SearchParamsType) => {
             await queryClient.invalidateQueries({
               queryKey: [RECENT_SEARCH_KEY],
             });
-            updateQueryParams('title', keyword);
+            updateQueryParams('title', keyword, setCurrentPage);
           },
         },
       );
     } else {
-      updateQueryParams('title', keyword);
+      updateQueryParams('title', keyword, setCurrentPage);
     }
   };
 
@@ -98,5 +79,6 @@ export const usePostList = (initialSearchParams: SearchParamsType) => {
     handleKeywordSearch,
     updateQueryParams,
     handleSearchClick,
+    setCurrentPage,
   };
 };
