+++ src/hooks/usePostListForTag.ts
@@ -1,40 +1,22 @@
-import { useCallback, useState } from 'react';
+import { useState } from 'react';
 import { useRouter } from 'next/navigation';
 import { SearchTagParamsType } from '@/types/posts.types';
+import useQueryParams from '@/hooks/useQueryParams';
 
 export const usePostListForTag = (initialSearchParams: SearchTagParamsType) => {
   const router = useRouter();
   const [currentPage, setCurrentPage] = useState<number>(1);
   const [activeTab, setActiveTab] = useState<string>('all');
-
-  const updateQueryParams = useCallback(
-    (key: string, value: string | number | undefined) => {
-      const query = new URLSearchParams(window.location.search);
-
-      if (value && !(key === 'tutorId' && value === 'all')) {
-        query.set(key, String(value));
-      } else {
-        query.delete(key);
-      }
-
-      if (key !== 'page') {
-        setCurrentPage(1);
-        query.set('page', '1');
-      }
-
-      router.push(`/tag?${query.toString()}`);
-    },
-    [router],
-  );
+  const updateQueryParams = useQueryParams();
 
   const handleTabChange = (newTab: string) => {
     setActiveTab(newTab);
-    updateQueryParams('postType', newTab);
+    updateQueryParams('postType', newTab, setCurrentPage);
   };
 
   const handlePageChange = (page: number) => {
     setCurrentPage(page);
-    updateQueryParams('page', page);
+    updateQueryParams('page', page, setCurrentPage);
   };
 
   return {
@@ -42,6 +24,5 @@ export const usePostListForTag = (initialSearchParams: SearchTagParamsType) => {
     activeTab,
     handleTabChange,
     handlePageChange,
-    updateQueryParams,
   };
 };
